rearrange some structures
