+++ src/ds/mod.rs
@@ -1,11 +1,13 @@
 use super::*;
 
+mod thread_cache;
 mod dll;
 mod lru;
 mod radix;
 pub mod stack;
 
-pub use self::dll::Dll;
+use self::dll::Dll;
+pub use self::thread_cache::ThreadCache;
 pub use self::lru::Lru;
 pub use self::radix::Radix;
 pub use self::stack::{Stack, StackIter, node_from_frag_vec};
