+++ src/io/log/iterator.rs
